Create README.md
This Python script was written for converting Lisu text in non unicode formats into Unicode format for use on the web, mobile devices, or modern software. You can easily run this using Python launcher (Mac). Make sure the script and the files to be converted are in the same folder to remove any margin for errors.

Note: This does not support the Naxi language inverted "Y", which probably won't be a problem unless you're converting a document that originated in Yunnan.